latest development

+++ controllers/categoryController.js
@@ -3,23 +3,19 @@ const mongoose = require('mongoose');
 
 //GET all Categories
 const getCategories = async (req, res) => {
-  const categories = await Category.find({
-    userID: req.headers.jwt.userId,
-  }).sort({ createdAt: -1 });
+  const categories = await Category.find({userID: req.headers.jwt.userId}).sort({ createdAt: -1 });
 
   res.status(200).json(categories);
 };
 
-//GET Categories by type
+//GET Categories by ctype
 const getCategoriesByType = async (req, res) => {
-  const { type } = req.params;
+  const { ctype } = req.params;
 
-  const categories = await Category.find({ type: type }).sort({
-    createdAt: -1,
-  });
+  const categories = await Category.find({type: ctype});
 
   res.status(200).json(categories);
-};
+}
 
 //GET a single Category
 const getCategory = async (req, res) => {
@@ -34,25 +30,23 @@ const getCategory = async (req, res) => {
     return res.status(404).json({ error: 'No such Category' });
   }
 
-  if (category.userID != req.headers.jwt.userId)
-    //if category does not belong to the user
-    return res.status(404).json({ error: 'No such Category' });
+  if(category.userID != req.headers.jwt.userId) //if category does not belong to the user
+  return res.status(404).json({ error: 'No such Category' });
 
   res.status(200).json(category);
 };
 
 //POST a new Category
 const createCategory = async (req, res) => {
-  const { type, name, accountID, icon, color } = req.body;
+  const { ctype, cname, iconID, colorhex } = req.body;
   const userID = req.headers.jwt.userId;
   try {
     const category = await Category.create({
-      type,
-      name,
-      accountID,
+      ctype,
+      cname,
       userID,
-      icon,
-      color,
+      iconID,
+      colorhex,
     });
     res.status(200).json(category);
   } catch (error) {
@@ -67,10 +61,7 @@ const deleteCategory = async (req, res) => {
     return res.status(404).json({ error: 'No such Category' });
   }
 
-  const category = await Category.findOneAndDelete({
-    _id: id,
-    userID: req.headers.jwt.userId,
-  });
+  const category = await Category.findOneAndDelete({ _id: id, userID: req.headers.jwt.userId });
 
   if (!category) {
     return res.status(400).json({ error: 'No such Category' });
